Multivariate normal design specification

Stores information necessary to simulate and visualize datasets based on underlying distribution multivariate normal distribution Z.

Usage

simdesign_mvtnorm(
  relations_initial,
  mean_initial = 0,
  sd_initial = 1,
  is_correlation = TRUE,
  method = "svd",
  name = "Multivariate-normal based simulation design",
  ...
)

Arguments

relations_initial

Correlation / Covariance matrix of the initial multivariate Normal distribution Z.

mean_initial

Vector of mean values of the initial multivariate Normal distribution Z. Dimension needs to correspond to dimension of relations.

sd_initial

Vector of standard deviations of the initial multivariate Normal distribution Z. Dimension needs to correspond to dimension of relations. Overriden by suqare root of diagonal elements of relations if is_correlation is FALSE.

is_correlation

If TRUE, then relations specifies a correlation matrix (default, this type of specification is usually more natural than specifying a covariance matrix). Otherwise, relations specifies a covariance matrix whose square root diagonal elements override sd_initial.

method

method argument of mvtnorm::rmvnorm.

name

Character, optional name of the simulation design.

...

Further arguments are passed to the simdesign constructor.

Value

List object with class attribute "simdesign_mvtnorm" (S3 class), inheriting from "simdesign". It contains the same entries as a simdesign

object but in addition the following entries:

mean_initial

sd_initial

cor_initial

Initial correlation matrix of multivariate normal distribution

Details

This S3 class implements a simulation design based on an underlying multivariate normal distribution by creating a generator function based on mvtnorm::rmvnorm.

Note

Note that relations specifies the correlation / covariance of the underlying Normal data Z and thus does not directly translate into correlations between the variables of the final datamatrix X.

Data Generation

Data will be generated by simulate_data using the following procedure:

The underlying data matrix Z is sampled from a multivariate Normal distribution (number of dimensions specified by dimensions of relations).

Z is then transformed into the final dataset X by applying the transform_initial function to Z.

X is post-processed if specified.

See also